// src/api/generated/model/galleryControllerGetGallerySort.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Honest Bangkok API
 * The Honest Bangkok API for locals and tourists to share tips, prices, and scam alerts.
 * OpenAPI spec version: 1.0
 */

export type GalleryControllerGetGallerySort = typeof GalleryControllerGetGallerySort[keyof typeof GalleryControllerGetGallerySort];


export const GalleryControllerGetGallerySort = {
  newest: 'newest',
  popular: 'popular',
} as const;